test(keys): close round-2 prover and architect findings (#1056)
- The rebinding-body guard rebinds ONCE (a per-fire rebind
  re-enters deactivate's cancel on the tick and rescued the
  pre-fix shape — prover round 2), asserts fail-closed, and
  pins that the tick never re-fires the body.
- The dispatch-entry seal walks the test trees too, since
  dispatchCommand went module-internal in the file split.
- The cue funnel is proved per function rather than by a
  balance identity two offsetting edits could satisfy.
- HotkeyRegistrar carries the never-fire-a-failed-registration
  contract where implementors read it; installHandlerIfNeeded
  documents the construction-is-inert dependency the seam
  suite's bare CarbonHotkeyCenter() leans on; ReleaseRegistrar
  splits out of the wiring suite at the file ceiling.
